pyproject.toml
[project]
name = "virtualfit-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aerich>=0.9.1",
    "argon2-cffi>=25.1.0",
    "fastapi[standard]>=0.116.2",
    "gunicorn>=23.0.0",
    "hf-xet>=1.1.10",
    "httpx>=0.28.1",
    "huggingface-hub[cli,hf-xet]>=0.35.1",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "opencv-python>=4.11.0.86",
    "orjson>=3.11.3",
    "pillow>=11.3.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
    "torchgeometry>=0.1.2",
    "torchvision>=0.23.0",
    "tortoise-orm[asyncpg]>=0.25.1",
]
[project.scripts]
dev = "src.scripts.server:run_server"
prod = "src.scripts.server:run_prod"
makemigrations = "src.scripts.aerich:makemigrations"
migrate = "src.scripts.aerich:migrate"
seed = "src.scripts.seed:run"

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "src",

]

[tool.aerich]
tortoise_orm = "src.core.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
